HSEARCH-4563 Unify query/simple query predicate steps
