+++ engine/src/main/java/org/hibernate/search/engine/search/predicate/dsl/QueryStringPredicateFieldMoreStep.java
@@ -7,8 +7,8 @@
 package org.hibernate.search.engine.search.predicate.dsl;
 
 /**
- * The step in a "simple query string" predicate definition where the query string to match can be set
- * (see the superinterface {@link SimpleQueryStringPredicateMatchingStep}),
+ * The step in a "query string" predicate definition where the query string to match can be set
+ * (see the superinterface {@link QueryStringPredicateMatchingStep}),
  * or optional parameters for the last targeted field(s) can be set,
  * or more target fields can be added.
  *
@@ -18,40 +18,6 @@ package org.hibernate.search.engine.search.predicate.dsl;
 public interface QueryStringPredicateFieldMoreStep<
 		S extends QueryStringPredicateFieldMoreStep<?, N>,
 		N extends QueryStringPredicateOptionsStep<?>>
-		extends QueryStringPredicateMatchingStep<N>, MultiFieldPredicateFieldBoostStep<S> {
-
-	/**
-	 * Target the given field in the simple query string predicate,
-	 * as an alternative to the already-targeted fields.
-	 * <p>
-	 * Only text fields are supported.
-	 * <p>
-	 * See {@link SimpleQueryStringPredicateFieldStep#field(String)} for more information on targeted fields.
-	 *
-	 * @param fieldPath The <a href="SearchPredicateFactory.html#field-paths">path</a> to the index field
-	 * to apply the predicate on.
-	 * @return The next step.
-	 *
-	 * @see SimpleQueryStringPredicateFieldStep#field(String)
-	 */
-	default S field(String fieldPath) {
-		return fields( fieldPath );
-	}
-
-	/**
-	 * Target the given fields in the simple query string predicate,
-	 * as an alternative to the already-targeted fields.
-	 * <p>
-	 * Only text fields are supported.
-	 * <p>
-	 * See {@link SimpleQueryStringPredicateFieldStep#fields(String...)} for more information on targeted fields.
-	 *
-	 * @param fieldPaths The <a href="SearchPredicateFactory.html#field-paths">paths</a> to the index fields
-	 * to apply the predicate on.
-	 * @return The next step.
-	 *
-	 * @see SimpleQueryStringPredicateFieldStep#fields(String...)
-	 */
-	S fields(String... fieldPaths);
+		extends CommonQueryStringPredicateFieldMoreStep<S, N> {
 
 }
