+++ engine/src/main/java/org/hibernate/search/engine/search/predicate/dsl/SimpleQueryStringPredicateFieldStep.java
@@ -11,44 +11,8 @@ package org.hibernate.search.engine.search.predicate.dsl;
  *
  * @param <N> The type of the next step.
  */
-public interface SimpleQueryStringPredicateFieldStep<N extends SimpleQueryStringPredicateFieldMoreStep<?, ?>> {
+public interface SimpleQueryStringPredicateFieldStep<N extends SimpleQueryStringPredicateFieldMoreStep<?, ?>>
+		extends CommonQueryStringPredicateFieldStep<N> {
 
-	/**
-	 * Target the given field in the simple query string predicate.
-	 * <p>
-	 * Only text fields are supported.
-	 * <p>
-	 * Multiple fields may be targeted by the same predicate:
-	 * the predicate will match if <em>any</em> targeted field matches.
-	 * <p>
-	 * When targeting multiple fields, those fields must have compatible types.
-	 * Please refer to the reference documentation for more information.
-	 *
-	 * @param fieldPath The <a href="SearchPredicateFactory.html#field-paths">path</a> to the index field
-	 * to apply the predicate on.
-	 * @return The next step.
-	 */
-	default N field(String fieldPath) {
-		return fields( fieldPath );
-	}
-
-	/**
-	 * Target the given fields in the simple query string predicate.
-	 * <p>
-	 * Only text fields are supported.
-	 * <p>
-	 * Equivalent to {@link #field(String)} followed by multiple calls to
-	 * {@link SimpleQueryStringPredicateFieldMoreStep#field(String)},
-	 * the only difference being that calls to {@link SimpleQueryStringPredicateFieldMoreStep#boost(float)}
-	 * and other field-specific settings on the returned step will only need to be done once
-	 * and will apply to all the fields passed to this method.
-	 *
-	 * @param fieldPaths The <a href="SearchPredicateFactory.html#field-paths">paths</a> to the index fields
-	 * to apply the predicate on.
-	 * @return The next step.
-	 *
-	 * @see #field(String)
-	 */
-	N fields(String... fieldPaths);
 
 }
