+++ engine/src/main/java/org/hibernate/search/engine/search/predicate/dsl/SimpleQueryStringPredicateOptionsStep.java
@@ -10,45 +10,13 @@ import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.Set;
 
-import org.hibernate.search.engine.search.common.BooleanOperator;
-
 /**
- * The final step in an "simple query string" predicate definition, where optional parameters can be set.
+ * The final step in a "simple query string" predicate definition, where optional parameters can be set.
  *
  * @param <S> The "self" type (the actual exposed type of this step).
  */
 public interface SimpleQueryStringPredicateOptionsStep<S extends SimpleQueryStringPredicateOptionsStep<?>>
-		extends PredicateFinalStep, PredicateScoreStep<S> {
-
-	/**
-	 * Define the default operator.
-	 * <p>
-	 * By default, unless the query string contains explicit operators,
-	 * documents will match if <em>any</em> term mentioned in the query string is present in the document ({@code OR} operator).
-	 * This can be used to change the default behavior to {@code AND},
-	 * making document match if <em>all</em> terms mentioned in the query string are present in the document.
-	 *
-	 * @param operator The default operator ({@code OR} or {@code AND}).
-	 * @return {@code this}, for method chaining.
-	 */
-	S defaultOperator(BooleanOperator operator);
-
-	/**
-	 * Define an analyzer to use at query time to interpret the value to match.
-	 * <p>
-	 * If this method is not called, the analyzer defined on the field will be used.
-	 *
-	 * @param analyzerName The name of the analyzer to use in the query for this predicate.
-	 * @return {@code this}, for method chaining.
-	 */
-	S analyzer(String analyzerName);
-
-	/**
-	 * Any analyzer or normalizer defined on any field will be ignored to interpret the value to match.
-	 *
-	 * @return {@code this}, for method chaining.
-	 */
-	S skipAnalysis();
+		extends CommonQueryStringPredicateOptionsStep<S> {
 
 	/**
 	 * Enable operation in the given flags.
